refactor: 컨트롤러와 서비스에서의 인자 로깅 방법 수정
